✅ test(i18n): Step D-7 — schema-invariant rename of sourceId scaffold
Renames `PastResultsLanguageCompatTests` → `BundledPresetSourceIdSchemaTests`
to reflect the honest scope: the activated test asserts the
ADR-010 D4 schema invariant (JA + EN bundled-preset siblings share a
canonical `sourceId`), NOT a Past Results UI consumer that aggregates
across language variants — that consumer doesn't ship in Step D.

Changes:
- File renamed; suite + test renamed accordingly (`perLanguagePresetsShareSourceIdAfterStepD`
  → `perLanguagePresetsShareCanonicalSourceId`).
- `.disabled(...)` trait removed — the activation criteria (PresetLoader
  writes canonical sourceId on bundled inserts) are satisfied by
  Step D-3 (#388 item 3).
- Test body strengthened: in addition to count-based aggregation,
  the assertion now pins the exact id set returned by sourceId
  filtering — locks the data shape, not just the cardinality.
- Docstring rewritten: schema-only scope, history note (rename
  rationale), explicit pointer to the follow-up consumer-surface
  tracking issue.
- New tracking issue #392 ("Past Results cross-language aggregation
  consumer") created — captures the ResultsView / SimulationRepository
  wiring deferred from Step D, plus UX decisions for per-variant
  labelling.

`scripts/xcodebuild.sh test -only-testing PasturaTests/BundledPresetSourceIdSchemaTests`:
1 test passed.

See #388 — Item 7 of 8. Follow-up: #392.
